requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
llvmlite==0.43.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.0
openai-whisper==20231117
packaging==24.1
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.27.2
PyAudio==0.2.14
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
simpleaudio==1.0.4
sympy==1.12.1
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
